refactor(drone_delivery): remove _prefix wrapper pattern from bt_nodes and consolidate scenario sim files into sim/ subfolders
